Ejercicio 09 -> Cambios en los nombres de los controles agregados al formulario de Agregar Cuerpos Celestes y funciones para cargar datos. Leves cambios en la visibilidad de metodos dentro de la clase Observatorio.

+++ EJERCICIO_09/OBSERVATORIO/clases/CuerpoCeleste.cs
@@ -15,8 +15,8 @@ namespace OBSERVATORIO.clases
 			set { nombre = value; }
 		}
 
-		private int edad;
-		public int Edad
+		private float edad;
+		public float Edad
 		{
 			get { return edad; }
 			set { edad = value; }
